Распределяет код по модулям, создает точку входа
