+++ js/function.js
@@ -32,3 +32,20 @@ const getOptimalStringWithPad = (str, minLength, pad) => { /* С помощью 
 };
 
 getOptimalStringWithPad('Ha', 4, 'Ho');
+
+const getRandomInteger = (a, b) => { /* Возвращает случайное целое число в заданном диапазоне */
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)]; /* Возвращает случайный элемент массива */
+
+export {isStringOptimal};
+export {isItPalindrome};
+export {getNumber};
+export {getOptimalStringWithPad};
+export {getRandomArrayElement};
+export {getRandomInteger};
